+++ FEstimatorLines.h
@@ -8,7 +8,7 @@
 
 struct lineCorrespStruct {
     //cv::line_descriptor::KeyLine line1, line2;
-    float line1Angle, line2Angle, line1Length, line2Length;
+    double line1Angle, line2Angle, line1Length, line2Length;
     Mat line1StartNormalized, line1EndNormalized, line2StartNormalized, line2EndNormalized;
     Mat line1Start, line1End, line2Start, line2End;
     int id;
@@ -16,13 +16,13 @@ struct lineCorrespStruct {
 
 struct lineCorrespSubsetError {
     int lineCorrespIdx;
-    float lineCorrespError;
+    double lineCorrespError;
 };
 
 struct lineSubsetStruct {
     std::vector<lineCorrespStruct> lineCorrespondencies;
     //std::vector<int> lineCorrespondenceIdx;
-    float qualityMeasure;
+    double qualityMeasure;
     double meanSquaredSymmeticTransferError;
     Mat Hs, Hs_normalized;
     //std::vector<lineCorrespSubsetError> correspErrors;
@@ -39,24 +39,24 @@ public:
 
 private:
 
-    float levenbergMarquardt(lineSubsetStruct &bestSubset);
+    double levenbergMarquardt(lineSubsetStruct &bestSubset);
     bool computeHomography(lineSubsetStruct &subset);
     double squaredSymmeticTransferError(Mat H_invT, Mat H_T, lineCorrespStruct lc);
     double squaredSymmeticTransferError(Mat H, lineCorrespStruct lc);
     void visualizeProjectedLines(lineSubsetStruct subset, int lineWidth, bool drawConnections, std::string name);
     bool hasGeneralPosition(std::vector<int> subsetsIdx, int newIdx, std::vector<lineCorrespStruct> lineCorrespondencies);
     lineSubsetStruct calcRANSAC(std::vector<lineSubsetStruct> &subsets, double threshold, std::vector<lineCorrespStruct> lineCorrespondencies);
-    int filterLineExtractions(float minLenght, std::vector<cv::line_descriptor::KeyLine> &keylines);
+    int filterLineExtractions(double minLenght, std::vector<cv::line_descriptor::KeyLine> &keylines);
     //int filterLineMatches(std::vector<DMatch> &matches);
     void fillHLinEq(Mat &linEq, std::vector<lineCorrespStruct> correspondencies);
-    void fillHLinEqBase(Mat &linEq, float x, float y, float A, float B, float C, int row);
+    void fillHLinEqBase(Mat &linEq, double x, double y, double A, double B, double C, int row);
     lineSubsetStruct calcLMedS(std::vector<lineSubsetStruct> &subsets, std::vector<lineCorrespStruct> lineCorrespondencies);
-    float calcMedS(lineSubsetStruct &subset, std::vector<lineCorrespStruct> lineCorrespondencies);
+    double calcMedS(lineSubsetStruct &subset, std::vector<lineCorrespStruct> lineCorrespondencies);
     Mat* normalizeLines(std::vector<lineCorrespStruct> &correspondencies);
     void filterUsedLineMatches(std::vector<lineCorrespStruct> &matches, std::vector<lineCorrespStruct> usedMatches);
-    bool findHomography(std::vector<lineCorrespStruct> &lineCorrespondencies, int method, float confidence, lineSubsetStruct &result);
+    bool findHomography(std::vector<lineCorrespStruct> &lineCorrespondencies, int method, double confidence, lineSubsetStruct &result);
     lineSubsetStruct estimateHomography(std::vector<lineCorrespStruct> lineCorrespondencies, int method, int sets);
-    lineCorrespStruct getlineCorrespStruct(float start1x, float start1y, float end1x, float end1y, float start2x, float start2y, float end2x, float end2y, int id);
+    lineCorrespStruct getlineCorrespStruct(double start1x, double start1y, double end1x, double end1y, double start2x, double start2y, double end2x, double end2y, int id);
     lineCorrespStruct getlineCorrespStruct(cv::line_descriptor::KeyLine l1, cv::line_descriptor::KeyLine l2, int id);
     lineCorrespStruct getlineCorrespStruct(lineCorrespStruct lcCopy);
     void visualizeMatches(std::vector<lineCorrespStruct> correspondencies, int lineWidth, bool drawConnections, std::string name);
