+++ FEstimatorPoints.cpp
@@ -9,8 +9,8 @@ FEstimatorPoints::FEstimatorPoints(Mat img1, Mat img2, Mat img1_c, Mat img2_c, s
     if(LOG_DEBUG) std::cout << "Estimating: " << name << std::endl;
     successful = false;
 
-    normT1 = Mat::eye(3,3,CV_32FC1);
-    normT2 = Mat::eye(3,3,CV_32FC1);
+    normT1 = Mat::eye(3,3,CV_64FC1);
+    normT2 = Mat::eye(3,3,CV_64FC1);
 }
 
 int FEstimatorPoints::extractMatches() {
@@ -85,7 +85,7 @@ bool FEstimatorPoints::compute() {
     int used = 0;
     extractMatches();
     F = findFundamentalMat(x1, x2, FM_RANSAC, 3.0, 0.99, mask);
-    F.convertTo(F, CV_32FC1);
+    F.convertTo(F, CV_64FC1);
     for(int i = 0; i < x1.size(); i++) {
         if(mask.at(i)) {
             x1_used.push_back(x1.at(i));
@@ -108,15 +108,15 @@ bool FEstimatorPoints::compute() {
     return true;
 }
 
-std::vector<Point2f> FEstimatorPoints::getX1() {
+std::vector<Point2d> FEstimatorPoints::getX1() {
     return x1_used;
 }
 
-std::vector<Point2f> FEstimatorPoints::getX2() {
+std::vector<Point2d> FEstimatorPoints::getX2() {
     return x2_used;
 }
 
-void FEstimatorPoints::visualizeMatches(std::vector<Point2f> p1, std::vector<Point2f> p2, int lineWidth, bool drawConnections, std::string name) {
+void FEstimatorPoints::visualizeMatches(std::vector<Point2d> p1, std::vector<Point2d> p2, int lineWidth, bool drawConnections, std::string name) {
     Mat img;
     hconcat(image_1_color.clone(), image_2_color.clone(), img);
     for(int i = 0; i < p1.size(); i++) {
